make gargantext buildable with stack & cabal in a nix environment
I revamped the whole Nix infrastructure a bit to clearly delineate the bits
about providing a suitable environment (with respect to non Haskell deps) and
the shell and stack helper expressions, and moved it all under `nix/` to
make it easier to ignore (if not interested) or focus on (if interested) the Nix
infrastructure of this repository.

For some reason, this required changes that don't seem necessary in the
Docker environment, even though this all happens using the same GHC version
as used in the Docker environment/Stackage LTS snapshot.

The said changes correspond to the source repositories mentioned in the
cabal.project and stack.yaml files that have not been left untouched, for which
I either pointed to a fork I made, or a different commit that I pushed in a
dedicated branch (for the ISCPIF gitlab repositories to which I seemingly
have write access).

Many of the changes were about Haddock `-- |` style syntax being used in an
unexpected place, and this was making GHC's parser quite sad.

This commit doesn't (yet?) provide Nix infrastructure for building everything
in/through Nix, but shouldn't be too much work with something like haskell.nix,
which has suitable tooling to import stack.yaml or cabal.project based
project specifications and turn them into fully nixified builds, which might
or might not be desirable.